feat: migrate to linkd for dependency injection
